Adding routes and clickable pages
Imported routing from react-router-dom. and displaying the page underneath the navigation bar.
The pages the names are really just placeholders which should be changed. Added the home page which is also the profile page (I think). Added the image display page which I am working on. I just did this to set up the pages as simply as possible without any css. Next, I might change the href to link so the page does not reload each time you click on a link.