Removed Home, Added Login
I commented all the code on the home page since it doesn't run at the moment. If the home branch need to merge, I might have to add the code back in to resolve conflicts. I added the code from the login branch to the login page just to see. If we merge the login branch, there are probably going to be some conflicts. Added some more place holder stuff to the display page.