Connected Backend to Frontend
I changed some lines of code in the backend so that the routes were properly imported. Then, I used a test.rest file to test out if the backend was working by sending different requests. I have only tried out registering, logging in, and getting posts/users. There was an error in post/service in which all I had to do was add new. Added a register page to the frontend which should work along with the login page (which I took from Montag's branch).  Im just testing out the fetching posts thing in the Image display page, though it seems to be fetching users instead of posts. It seems like to add query parameters to the fetch request, it had to be formatted like /?parameter=value or something like that. Anyway, the backend and frontend seem to be working okay.