Fix eslint compiler warning about logo and revert Display link
I found why Routes are important... They're just kind of like a variable name to redirect the path of a local file to a "name" which can be referred elsewhere. I think it's useful for readability when scaling though it can technically just be replaced with the direct relative link to the source file.