Added Image uploading
I figured out how to upload images. Important thing is that we kinda have to upload an array since the backend relies on it. Also, multer has a very specific way of setting up the frontend so that the files actually get sent. I am not really sure what makes it work but when I change anything, things start to break. For multer, you need to make sure the file input name matches the name in the backend. Also, the form has encType="multipart/form-data" and we get the files from searching for the element by Id rather than our states. I couldn't figure out how to get axios to add the authorization header so I just did it the old way.

For now, image display shows all of your own posts(when you are logged in). We can probably make a profile page where you see all your information and posts and can edit everything.

I forget to say that I am keeping the authentication token in the local storage and just looking there when we need to authenticate.

I made a bunch of posts for testing that we probably need to delete. If you want to log into that profile to see the posts, the username is "abc@abc.com" and pass is "abcd". Probably will need to show a list of users and a way to see their profile. If we implement the search functionality, it also will help with seeing other user's profiles and their posts.