Added a User Profile
Clicking on the usernames on your 'For You' section of your profile sends you to the profile of that user.
I tried changing the names of the urls to better match the backend ones (post/:postId and user/:userId). Its important to get the names exactly the same since it doesnt work otherwise.
Nothing really is formatted correctly but it works for now.