Added commenting to posts
First I had to change all the url endpoints in the backend since many of them were using the same url.
When using navigate, if you want to replace the entire url rather than append to the end of it, use , { replace: true} and make sure your url starts with a /.
When making a comment the page refreshes to fetch all the comments again. There probably is a way to do this without refresh but I think its easier this way.