Added likes for posts
Same thing as comments. Fetch from react/:postId. This saves a list of all the likes including who liked. This might be a bit more than we need. When you click the like button, the next time you click it, you unlike. I guess its possible to show that you already liked it.
I am not sure if we are doing 'shares' since I do not see that in the backend.