Added Search and Pages for Userpage
I added search as a new page since it is easier to show a list of users. You can click on this list of users to go to their profiles. Searching nothing will give you every user.
I also added pages to the user profile so that you can view all of their posts if they exceed the per page limit of 10.
Pages could be used in other things like the profile/home.
There were a few changes in the backend, such as changing the urls, and adding more error checking.
I previously thought the getpost api was just getting a list of users but that actually was from the get user api which both used '/' as their url to fetch.