Nav bar changes when logged in
Also created a logout button on top left.
Logged in shows Upload, Profile, Search.
Not logged in shows Log In, Register
Completely rewrote app.js and navbar exactly like I did it for a previous project so that when you first visit the page and are not logged in you can only see pages that do not need authentication.

I think this might have been a problem if anyone had tried to visit the website and were not already logged in. Now, there should not be any errors when you are not logged in.

Also aligned the navigation bar to the center which took a bit of time.