Fixed some bugs with useContext
Previously, when running npm start, there would be a blank page until you reload. Now, we make sure to set the state and check if user exists before doing this like fetching the user's posts. Also, when logging in for the first time, the state is not set immediately so we might want to reload the page inorder to run the useEffect thing again and fetch the posts of the user.
I also changed getting the user profile info from fetching user/:my username to /me to avoid using username before it exists.