Adding abilty to follow / error checking
You can now click on a user's profile and follow them. Their posts now appear in your follow list on your profile and do not appear in the for you section.
There were some error for if you followed everyone and there were no posts left. So adding more error checking on the fetch requests. There are probably a lot of console.logs and the page reloads and fetches more times than needed but it seems to work.